Update menu on save, handle save from any window
- fixes #60
- fixes #61
